backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-generativeai
requests
beautifulsoup4
schedule